// apps/web/src/lib/api/generated/main/model/roleResponseDtoOutput.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { RoleResponseDtoOutputPermissionsItem } from './roleResponseDtoOutputPermissionsItem';

export interface RoleResponseDtoOutput {
  id: string;
  guildId: string;
  name: string;
  /** @nullable */
  color?: number | null;
  /** @nullable */
  position?: number | null;
  permissions: RoleResponseDtoOutputPermissionsItem[];
  /** @nullable */
  lvlRangeFrom?: number | null;
  /** @nullable */
  lvlRangeTo?: number | null;
}